Pull slice data in Config component itself
